move conftest.oy
